requirements.txt
robotframework==6.1.1
robotframework-seleniumlibrary==6.2.0
pyyaml==6.0.1
webdriver-manager==4.0.2
robotframework-datadriver==1.6.1
